The backend spring boot configuration has been initialized with the MongoDB configurations; new dependencies have been removed from the pom.xml file and an update to the .gitignore file have been made.
The spring.data.mongodb properties in the application.properties have been updated with the MongoDB URI and database information. This allows easy setup for the database connection.

In the pom.xml, the oauth2-client dependency has been removed because we've decided not to implement the OAuth2 for our security needs at this time.

A line was added within the .gitignore file to ignore the 'application-secret.properties' configuration file present under 'src/main/resources' directory in backend. This action was taken to ensure the confidential information within this ignored file is not pushed to the public repository.